+++ cpp/src/jni/parquet/connector.h
